requirements.txt
Django>=5.0.7,<5.1
djangorestframework>=3.15.2,<3.16
djangorestframework-simplejwt>=5.3.1,<5.4
psycopg2>=2.9.9,<3.0
drf-spectacular>=0.27.2,<0.28
Pillow>=10.4.0,<10.5.0
olefile>=0.46,<0.47
django-phonenumber-field[phonenumberslite]>=8.0.0,<8.1.0
django-filter>=24.3,<25.0
stripe>=11.1.1,<12.0.0
celery>=5.4.0,<5.5.0
redis>=5.2.0,<5.3.0